feat: add lock on execute sprite batch chore: renamed Native namespace to Runtimes fix: task node is not executing
