+++ REngine.Assets/Module.cs
@@ -12,8 +12,8 @@ namespace REngine.Assets
 	{
 		static AssetsModule()
 		{
-			NativeLibrary.SetDllImportResolver(typeof(SFML.Audio.Music).Assembly, Core.Native.NativeReferences.DefaultDllImportResolver);
-			NativeLibrary.SetDllImportResolver(typeof(FreeTypeSharp.NativeObject).Assembly, Core.Native.NativeReferences.DefaultDllImportResolver);
+			NativeLibrary.SetDllImportResolver(typeof(SFML.Audio.Music).Assembly, Core.Runtimes.NativeReferences.DefaultDllImportResolver);
+			NativeLibrary.SetDllImportResolver(typeof(FreeTypeSharp.NativeObject).Assembly, Core.Runtimes.NativeReferences.DefaultDllImportResolver);
 		}
 
 		public static void Setup(IServiceRegistry registry)
